Move particle texture loading to own file
